+++ huaweicloud/services/gaussdb/resource_huaweicloud_gaussdb_mysql_account_privilege.go
@@ -250,7 +250,7 @@ func resourceGaussDBAccountPrivilegeRead(_ context.Context, d *schema.ResourceDa
 			account = res
 			break
 		}
-		total := utils.PathSearch("total_count", getGaussDBAccountRespBody, 0).(float64)
+		total := utils.PathSearch("total_count", getGaussDBAccountRespBody, float64(0)).(float64)
 		currentTotal += pageNum
 		if currentTotal == int(total) {
 			break
