+++ huaweicloud/services/rds/resource_huaweicloud_rds_sql_audit.go
@@ -143,7 +143,7 @@ func resourceSQLAuditRead(_ context.Context, d *schema.ResourceData, meta interf
 		return common.CheckDeletedDiag(d, err, "error retrieving RDS SQL audit")
 	}
 
-	keepDays := utils.PathSearch("keep_days", getSQLAuditRespBody, 0).(float64)
+	keepDays := utils.PathSearch("keep_days", getSQLAuditRespBody, float64(0)).(float64)
 	if keepDays == 0 {
 		return common.CheckDeletedDiag(d, golangsdk.ErrDefault404{}, "")
 	}
@@ -282,7 +282,7 @@ func rdsSQLAuditStateRefreshFunc(client *golangsdk.ServiceClient, instanceID str
 		if err != nil {
 			return nil, "ERROR", err
 		}
-		keepDays := utils.PathSearch("keep_days", getSQLAuditRespBody, 0).(float64)
+		keepDays := utils.PathSearch("keep_days", getSQLAuditRespBody, float64(0)).(float64)
 		if keepDays == 0 {
 			return getSQLAuditRespBody, "DELETED", nil
 		}
